extract log table
